refactor: split bindings code of openssl into version-specific modules
